pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jpi-guard"
version = "0.1.0"
description = "Japanese Prompt Injection Guard — API client for jpi-guard (external-content-cleanse)"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "nexus-api-lab" }]
keywords = [
    "prompt-injection",
    "llm-security",
    "japanese",
    "langchain",
    "ai-safety",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27.0",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.2.0"]
llamaindex = ["llama-index-core>=0.10.0"]
all = ["langchain-core>=0.2.0", "llama-index-core>=0.10.0"]

[project.urls]
Homepage = "https://nexus-api-lab.com"
Repository = "https://github.com/nexus-api-lab/jpi-guard-python"
Documentation = "https://nexus-api-lab.com/docs"

[tool.hatch.build.targets.wheel]
packages = ["jpi_guard"]
